+++ app/jobs/rails_audit_log/write_audit_log_job.rb
@@ -1,20 +1,21 @@
 module RailsAuditLog
   class WriteAuditLogJob < ApplicationJob
-    def perform(entry_attrs, version_limit: nil)
+    def perform(entry_attrs, version_limit: nil, retention_period: nil)
       AuditLogEntry.create!(entry_attrs)
-
-      return unless version_limit
 
       item_type = entry_attrs["item_type"]
       item_id   = entry_attrs["item_id"]
-      count     = AuditLogEntry.where(item_type: item_type, item_id: item_id).count
-      excess    = count - version_limit
-      return unless excess > 0
+      scope     = AuditLogEntry.where(item_type: item_type, item_id: item_id)
 
-      AuditLogEntry.where(item_type: item_type, item_id: item_id)
-                   .order(id: :asc)
-                   .limit(excess)
-                   .delete_all
+      if retention_period
+        scope.where(created_at: ..retention_period.ago).delete_all
+      end
+
+      if version_limit
+        count  = scope.count
+        excess = count - version_limit
+        scope.order(id: :asc).limit(excess).delete_all if excess > 0
+      end
     end
   end
 end
